third_party/WebKit/LayoutTests/editing/style/create-block-for-style-010.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<p>
Should see the "foo" and "baz" lines centered in their own red boxes. The "baz" line should not be centered.
</p>
<div id="sample" contenteditable>foo<div>bar</div>baz</div>
<div id="log"></div>
<script>
test(function() {
    var selection = window.getSelection();
    var sample = document.getElementById('sample');

    selection.collapse(sample, 0);
    selection.extend(sample.childNodes[1].firstChild, 1);
    document.execCommand('justifyCenter');

    assert_equals(sample.innerHTML, '<div style="text-align: center;">foo</div><div style="text-align: center;">bar</div>baz', 'innerHTML');
    assert_false(selection.isCollapsed, 'isCollapsed');
    assert_equals(selection.anchorNode, sample.firstChild.firstChild, 'anchorNode');
    assert_equals(selection.anchorOffset, 0, 'anchorOffset');
    assert_equals(selection.focusNode, sample.childNodes[1].firstChild, 'focusNode');
    assert_equals(selection.focusOffset, 1, 'focusOffset');
});
</script>
